refactor op to symmetry
