+++ abismal/symmetry/__init__.py
@@ -1,1 +1,2 @@
 from .reciprocal_asu import ReciprocalASU,ReciprocalASUCollection
+from .op import Op
